pyproject.toml
[project]
name = "ai-evals-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Saurabh Pareek", email = "saurabh.pareek@valamis.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.85.0",
    "cryptography>=44.0.0",
    "fastapi>=0.135.1",
    "openai>=2.28.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[project.scripts]
start = "ai_evals_server.main:main"

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"
